requirements.txt
Django==3.2
psycopg2-binary>=2.9
django-allauth>=0.47
djangorestframework>=3.12
Pillow>=9.3.0